+++ controller/coq_lsp.ml
@@ -115,12 +115,7 @@ let add_message lvl loc msg q =
 
 let mk_fb_handler q Feedback.{ contents; _ } =
   match contents with
-  | Message (((Error | Warning | Notice) as lvl), loc, msg) ->
-    add_message lvl loc msg q
-  | Message ((Info as lvl), loc, msg) ->
-    if !Fleche.Config.v.show_coq_info_messages then add_message lvl loc msg q
-    else ()
-  | Message (Debug, _loc, _msg) -> ()
+  | Message (lvl, loc, msg) -> add_message lvl loc msg q
   | _ -> ()
 
 let coq_init ~fb_queue ~debug =